staging: unisys: change unsigned long to u32
This patch changes unsigned long to u32 in channel.h. This
change is needed in channel.h, because it needs to be size
specific in order to interact with Unisys s-Par firmware.

+++ drivers/staging/unisys/common-spar/include/channels/channel.h
@@ -319,8 +319,8 @@ spar_check_channel_client(void __iomem *ch,
 	}
 	if (expected_version > 0) {	/* caller wants us to verify
 					 * channel version */
-		unsigned long ver = readl(&((struct channel_header __iomem *)
-				    (ch))->version_id);
+		u32 ver = readl(&((struct channel_header __iomem *)
+				  (ch))->version_id);
 		if (ver != expected_version)
 			return 0;
 	}
